+++ service-worker.js
@@ -1,6 +1,6 @@
 "use strict";
 
-const CACHE_NAME = "driver-command-v3.3.0";
+const CACHE_NAME = "driver-command-v3.4.0";
 const APP_SHELL = [
   "./",
   "./index.html",